test(identifications): fold bulk endpoint tests into the main tests module
The bulk tests lived in a separate ami/main/test_bulk_identifications.py
module; identification tests in this app live in ami/main/tests.py, so the
classes move there unchanged apart from a renamed endpoint constant.
